PCAL pcal_int from updated difxio is float (and in type_308.h record is anyway always float not int)
git-svn-id: https://svn.atnf.csiro.au/difx@8026 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/difx2mark4/trunk/src/difx2mark4.h
@@ -110,7 +110,7 @@ struct fblock_tag
         int bs;                     // quantization bits/sample
         int first_time;             // true iff first entry in table of chan_id for ant 
         int zoom;                   // true iff this channel is zoom mode
-        int pcal_int;               // pcal interval (MHz)
+        double pcal_int;            // pcal interval (MHz)
         double freq;                // LO frequency (MHz); negative for LSB
         double bw;                  // bandwidth (MHz)
         } stn[2];                   // reference | remote
